Remove compilation switch for default use of right/left operators
To achieve consistent behavior, the compilation switch was removed as the differences of Positions (or all linear types) are always in the global frame. Therefore, per default, differences are computed using the left operator for framed units. To this end, the base change behavior of the difference units had to be adapted.

Additionally, the BaseChange struct now has a private constructor and can only be built using the make_base_change template function, as base changes are always computed using the right operators.